test_: split test steps into separate classes
