+++ tests-functional/steps/messenger.py
@@ -1,208 +1,18 @@
-from contextlib import contextmanager
-import json
 import logging
 import random
 import string
-import threading
 import time
-from collections import namedtuple
 from uuid import uuid4
-
 import pytest
 from tenacity import retry, stop_after_delay, wait_fixed
-from clients.services.wallet import WalletService
-from clients.signals import SignalClient, SignalType
-from clients.status_backend import RpcClient, StatusBackend
-from conftest import option
-from resources.constants import USE_IPV6, user_1, user_2
+from clients.signals import SignalType
+from clients.status_backend import StatusBackend
+from resources.constants import USE_IPV6
 from resources.enums import MessageContentType
+from steps.network_conditions import NetworkConditionsSteps
 
 
-class StatusDTestCase:
-    network_id = 31337
-
-    def setup_method(self):
-        self.rpc_client = RpcClient(option.rpc_url_statusd)
-
-
-class StatusBackendTestCase:
-
-    reuse_container = True  # Skip close_status_backend_containers cleanup
-    await_signals = [SignalType.NODE_LOGIN.value]
-
-    network_id = 31337
-
-    def setup_class(self):
-        self.rpc_client = StatusBackend(await_signals=self.await_signals)
-        self.wallet_service = WalletService(self.rpc_client)
-
-        self.rpc_client.init_status_backend()
-        self.rpc_client.restore_account_and_login()
-        self.rpc_client.wait_for_login()
-
-    def teardown_class(self):
-        for status_backend in option.status_backend_containers:
-            status_backend.container.stop(timeout=10)
-            option.status_backend_containers.remove(status_backend)
-            status_backend.container.remove()
-
-
-class WalletTestCase(StatusBackendTestCase):
-
-    def wallet_create_multi_transaction(self, **kwargs):
-        method = "wallet_createMultiTransaction"
-        transfer_tx_data = {
-            "data": "",
-            "from": user_1.address,
-            "gas": "0x5BBF",
-            "input": "",
-            "maxFeePerGas": "0xbcc0f04fd",
-            "maxPriorityFeePerGas": "0xbcc0f04fd",
-            "to": user_2.address,
-            "type": "0x02",
-            "value": "0x5af3107a4000",
-        }
-        for key, new_value in kwargs.items():
-            if key in transfer_tx_data:
-                transfer_tx_data[key] = new_value
-            else:
-                logging.info(f"Warning: The key '{key}' does not exist in the transferTx parameters and will be ignored.")
-        params = [
-            {
-                "fromAddress": user_1.address,
-                "fromAmount": "0x5af3107a4000",
-                "fromAsset": "ETH",
-                "type": 0,  # MultiTransactionSend
-                "toAddress": user_2.address,
-                "toAsset": "ETH",
-            },
-            [
-                {
-                    "bridgeName": "Transfer",
-                    "chainID": 31337,
-                    "transferTx": transfer_tx_data,
-                }
-            ],
-            f"{option.password}",
-        ]
-        return self.rpc_client.rpc_request(method, params)
-
-    def send_valid_multi_transaction(self, **kwargs):
-        response = self.wallet_create_multi_transaction(**kwargs)
-
-        tx_hash = None
-        self.rpc_client.verify_is_valid_json_rpc_response(response)
-        try:
-            tx_hash = response.json()["result"]["hashes"][str(self.network_id)][0]
-        except (KeyError, json.JSONDecodeError):
-            raise Exception(response.content)
-        return tx_hash
-
-
-class TransactionTestCase(WalletTestCase):
-
-    def setup_method(self):
-        self.tx_hash = self.send_valid_multi_transaction()
-
-
-class EthRpcTestCase(WalletTestCase):
-
-    @pytest.fixture(autouse=True, scope="class")
-    def tx_data(self):
-        tx_hash = self.send_valid_multi_transaction()
-        self.wait_until_tx_not_pending(tx_hash)
-
-        receipt = self.get_transaction_receipt(tx_hash)
-        try:
-            block_number = receipt.json()["result"]["blockNumber"]
-            block_hash = receipt.json()["result"]["blockHash"]
-        except (KeyError, json.JSONDecodeError):
-            raise Exception(receipt.content)
-
-        tx_data = namedtuple("TxData", ["tx_hash", "block_number", "block_hash"])
-        return tx_data(tx_hash, block_number, block_hash)
-
-    def get_block_header(self, block_number):
-        method = "ethclient_headerByNumber"
-        params = [self.network_id, block_number]
-        return self.rpc_client.rpc_valid_request(method, params)
-
-    def get_transaction_receipt(self, tx_hash):
-        method = "ethclient_transactionReceipt"
-        params = [self.network_id, tx_hash]
-        return self.rpc_client.rpc_valid_request(method, params)
-
-    def wait_until_tx_not_pending(self, tx_hash, timeout=10):
-        method = "ethclient_transactionByHash"
-        params = [self.network_id, tx_hash]
-        response = self.rpc_client.rpc_valid_request(method, params)
-
-        start_time = time.time()
-        while response.json()["result"]["isPending"] is True:
-            time_passed = time.time() - start_time
-            if time_passed >= timeout:
-                raise TimeoutError(f"Tx {tx_hash} is still pending after {timeout} seconds")
-            time.sleep(0.5)
-            response = self.rpc_client.rpc_valid_request(method, params)
-        return response.json()["result"]["tx"]
-
-
-class SignalTestCase(StatusDTestCase):
-    await_signals = []
-
-    def setup_method(self):
-        super().setup_method()
-        self.signal_client = SignalClient(option.ws_url_statusd, self.await_signals)
-
-        websocket_thread = threading.Thread(target=self.signal_client._connect)
-        websocket_thread.daemon = True
-        websocket_thread.start()
-
-
-class NetworkConditionTestCase:
-
-    @contextmanager
-    def add_latency(self, node, latency=300, jitter=50):
-        logging.info("Entering context manager: add_latency")
-        node.container_exec(f"apk add iproute2 && tc qdisc add dev eth0 root netem delay {latency}ms {jitter}ms distribution normal")
-        try:
-            yield
-        finally:
-            logging.info("Exiting context manager: add_latency")
-            node.container_exec("tc qdisc del dev eth0 root")
-
-    @contextmanager
-    def add_packet_loss(self, node, packet_loss=2):
-        logging.info("Entering context manager: add_packet_loss")
-        node.container_exec(f"apk add iproute2 && tc qdisc add dev eth0 root netem loss {packet_loss}%")
-        try:
-            yield
-        finally:
-            logging.info("Exiting context manager: add_packet_loss")
-            node.container_exec("tc qdisc del dev eth0 root netem")
-
-    @contextmanager
-    def add_low_bandwith(self, node, rate="1mbit", burst="32kbit", limit="12500"):
-        logging.info("Entering context manager: add_low_bandwith")
-        node.container_exec(f"apk add iproute2 && tc qdisc add dev eth0 root tbf rate {rate} burst {burst} limit {limit}")
-        try:
-            yield
-        finally:
-            logging.info("Exiting context manager: add_low_bandwith")
-            node.container_exec("tc qdisc del dev eth0 root")
-
-    @contextmanager
-    def node_pause(self, node):
-        logging.info("Entering context manager: node_pause")
-        node.container_pause()
-        try:
-            yield
-        finally:
-            logging.info("Exiting context manager: node_pause")
-            node.container_unpause()
-
-
-class MessengerTestCase(NetworkConditionTestCase):
+class MessengerSteps(NetworkConditionsSteps):
 
     await_signals = [
         SignalType.MESSAGES_NEW.value,
