+++ tests-functional/steps/wallet.py
@@ -1,0 +1,57 @@
+import json
+import logging
+from conftest import option
+from resources.constants import user_1, user_2
+from steps.status_backend import StatusBackendSteps
+
+
+class WalletSteps(StatusBackendSteps):
+
+    def wallet_create_multi_transaction(self, **kwargs):
+        method = "wallet_createMultiTransaction"
+        transfer_tx_data = {
+            "data": "",
+            "from": user_1.address,
+            "gas": "0x5BBF",
+            "input": "",
+            "maxFeePerGas": "0xbcc0f04fd",
+            "maxPriorityFeePerGas": "0xbcc0f04fd",
+            "to": user_2.address,
+            "type": "0x02",
+            "value": "0x5af3107a4000",
+        }
+        for key, new_value in kwargs.items():
+            if key in transfer_tx_data:
+                transfer_tx_data[key] = new_value
+            else:
+                logging.info(f"Warning: The key '{key}' does not exist in the transferTx parameters and will be ignored.")
+        params = [
+            {
+                "fromAddress": user_1.address,
+                "fromAmount": "0x5af3107a4000",
+                "fromAsset": "ETH",
+                "type": 0,  # MultiTransactionSend
+                "toAddress": user_2.address,
+                "toAsset": "ETH",
+            },
+            [
+                {
+                    "bridgeName": "Transfer",
+                    "chainID": 31337,
+                    "transferTx": transfer_tx_data,
+                }
+            ],
+            f"{option.password}",
+        ]
+        return self.rpc_client.rpc_request(method, params)
+
+    def send_valid_multi_transaction(self, **kwargs):
+        response = self.wallet_create_multi_transaction(**kwargs)
+
+        tx_hash = None
+        self.rpc_client.verify_is_valid_json_rpc_response(response)
+        try:
+            tx_hash = response.json()["result"]["hashes"][str(self.network_id)][0]
+        except (KeyError, json.JSONDecodeError):
+            raise Exception(response.content)
+        return tx_hash
