+++ tests-functional/tests/reliability/test_private_groups_messages.py
@@ -1,14 +1,14 @@
 from time import sleep
 from uuid import uuid4
 import pytest
-from tests.test_cases import MessengerTestCase
+from steps.messenger import MessengerSteps
 from clients.signals import SignalType
 from resources.constants import USE_IPV6
 
 
 @pytest.mark.usefixtures("setup_two_privileged_nodes")
 @pytest.mark.reliability
-class TestPrivateGroupMessages(MessengerTestCase):
+class TestPrivateGroupMessages(MessengerSteps):
 
     def test_private_group_messages_baseline(self, message_count=1):
         self.make_contacts()
