+++ tests-functional/tests/test_logging.py
@@ -1,5 +1,5 @@
 import re
-from test_cases import StatusBackend
+from clients.status_backend import StatusBackend
 import pytest
